Moved content of fastoad.base into fastoad fastoad.model_base
