+++ src/fastoad/model_base/__init__.py
@@ -1,3 +1,7 @@
+"""
+Base features for FAST-OAD models
+"""
+
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
 #  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
@@ -10,3 +14,6 @@
 #  GNU General Public License for more details.
 #  You should have received a copy of the GNU General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+from .atmosphere import Atmosphere, AtmosphereSI
+from .flight_point import FlightPoint
