+++ src/fastoad/model_base/tests/__init__.py
@@ -10,6 +10,3 @@
 #  GNU General Public License for more details.
 #  You should have received a copy of the GNU General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-from .atmosphere import Atmosphere, AtmosphereSI
-from .flight_point import FlightPoint
