+++ src/fastoad/model_base/tests/test_flight_point.py
