+++ src/fastoad/models/aerodynamics/aerodynamics_landing.py
@@ -17,7 +17,7 @@ Aero computation for landing phase
 import numpy as np
 import openmdao.api as om
 
-from fastoad.base import Atmosphere
+from fastoad.model_base import Atmosphere
 from fastoad.model_base.options import OpenMdaoOptionDispatcherGroup
 from .components.compute_max_cl_landing import ComputeMaxClLanding
 from .components.high_lift_aero import ComputeDeltaHighLift
