+++ src/fastoad/models/handling_qualities/tail_sizing/compute_ht_area.py
@@ -18,7 +18,7 @@ import numpy as np
 import openmdao.api as om
 from scipy.constants import g
 
-from fastoad.base import Atmosphere
+from fastoad.model_base import Atmosphere
 
 
 class ComputeHTArea(om.ExplicitComponent):
