+++ src/fastoad/models/performances/mission/openmdao/tests/test_mission.py
@@ -23,8 +23,8 @@ from numpy.testing import assert_allclose
 from openmdao.core.component import Component
 from scipy.constants import foot, knot
 
-from fastoad.base import FlightPoint
 from fastoad.io import VariableIO
+from fastoad.model_base import FlightPoint
 from fastoad.model_base.propulsion import AbstractFuelPropulsion, IOMPropulsionWrapper, IPropulsion
 from fastoad.module_management.service_registry import RegisterPropulsion
 from tests.testing_utilities import run_system
