+++ src/fastoad/models/performances/mission/segments/altitude_change.py
@@ -20,7 +20,7 @@ from typing import List, Tuple
 import pandas as pd
 from scipy.constants import foot, g
 
-from fastoad.base import AtmosphereSI, FlightPoint
+from fastoad.model_base import AtmosphereSI, FlightPoint
 from .base import ManualThrustSegment
 from ..exceptions import FastFlightSegmentIncompleteFlightPoint
 from ..util import get_closest_flight_level
